requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
groq
tenacity
pydantic
boto3
tweepy
instagrapi
streamlit
beautifulsoup4
lxml
rake-nltk
nltk
google-search-results
python-docx
tldextract
python-dateutil
dnspython
pyjwt
bcrypt
apscheduler
stable-baselines3
gym
markdown2
neo4j>=6.1.0
langsmith
numpy>=1.23.5,<2.3.0
scipy
chromadb
sentence-transformers